python-api/requirements.txt
Django==3.0.5
django-simple-history==2.12.0
djangorestframework==3.12.2
djongo==1.3.4
elastic-apm==6.0.0